+++ test/test-utils/expectations.ts
@@ -1,0 +1,5 @@
+export function expectToBeDefined<Element>(
+  element: Element | undefined,
+): asserts element is Element {
+  expect(element).toBeDefined();
+}
